backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
drf-spectacular==0.27.2
psycopg2-binary==2.9.11
python-dotenv==1.1.1
gunicorn==23.0.0
PyJWT==2.9.0
